--- a/gb-bus/src/area.rs
+++ b/gb-bus/src/area.rs
@@ -1,4 +1,4 @@
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash, PartialOrd, Ord)]
 pub enum Area {
     Rom,
     Vram,