--- a/gb-ppu/src/drawing/pixel.rs
+++ b/gb-ppu/src/drawing/pixel.rs
@@ -21,12 +21,12 @@ impl Pixel {
         }
     }
 
-    pub fn mix(&mut self, other: Pixel) {
+    pub fn mix(&mut self, other: &Pixel) {
         if !self.palette.get().is_sprite()
             && (!other.background_priority && other.palette.get().is_sprite() && other.color != 0)
             || (other.background_priority && other.palette.get().is_sprite() && self.color != 0)
         {
-            *self = other;
+            *self = other.clone();
         }
     }
 }