--- a/src/main.rs
+++ b/src/main.rs
@@ -57,17 +57,17 @@ fn main() {
             .start_frame()
             .expect("Fail at the start for the main window");
         if let Some(ref mut game) = game {
-            #[cfg(features = "time_frame")]
+            #[cfg(feature = "time_frame")]
             use std::time::Instant;
 
-            #[cfg(features = "time_frame")]
+            #[cfg(feature = "time_frame")]
             let now = Instant::now();
             while game.cycle() {
                 if let Some(flow) = debugger.updated_flow_status(game) {
                     game.update_scheduled_stop(flow);
                 }
             }
-            #[cfg(features = "time_frame")]
+            #[cfg(feature = "time_frame")]
             {
                 let elapsed = now.elapsed();
                 log::info!("frame ready in {}ms", elapsed.as_millis());