with these changes, I was able to reduce the `time_frame` 
- from `~56ms` to `~44ms` in `development`
- from `~4ms` to `~2ms` in `release`

both where compiled with the feature `time_frame` and the `log level` was set to `info`

depends on #367